Make CI green, and it found two crashes and two broken features doing it
The lint job was failing on 209 findings. Two of them were faults that would
bite a user, and neither had shown up in any of my own testing:

  * `keycapture.py` calls GLib.timeout_add() and never imports GLib. Recording
    a keyboard shortcut in the settings window ends in a NameError at the very
    last step — the same class of bug as the missing imports in window.py, and
    in a file I had not touched since.

  * `gen-defaults-doc.py` had `return "" if found.returncode else ""`. Both
    branches are the same empty string, so the "is it installed" column in
    DEFAULTS.md has always been blank and the function did nothing. It also
    shelled out to `command -v`, which is a shell builtin and not a program;
    shutil.which() is what it wanted.

Two more came out of running the settings window and reading its warnings:

  * an application called "OpenJDK 26 … Monitoring & Management Console" broke
    the applications list — Adw rows parse their titles as Pango markup, and an
    unescaped ampersand makes GTK drop the whole label. These names come from
    .desktop files nobody here controls, so every such title is escaped now.

  * my new GTK4 settings.ini set gtk-application-prefer-dark-theme, which
    libadwaita warns about on every application start and tells you to use
    color-scheme instead — which apply_gsettings() already does.

The rest was mechanical: 183 unused imports, stale noqa directives and unsorted
import blocks, all left over from splitting the settings app into modules and
copying its import header into every page.

Two config fixes rather than suppressions: `_` is declared as a builtin,
because gettext.install() genuinely puts it there and ruff was reporting F821
on a hundred correct lines; and panel/ and theme/ join the E402 exemption that
settings-gui/ and scripts/ already had, because gi.require_version() must run
before the import it guards.

Also regenerated KEYBINDS.md, which the four new shortcuts had made stale.

Verified: ruff clean, every CI step reproduced locally, all seven popups still
open, and the settings window now starts with no warnings at all.

+++ panel/osd_popup.py
@@ -24,9 +24,8 @@ import gi
 
 gi.require_version("Gtk", "4.0")
 
-from gi.repository import GLib, Gtk  # noqa: E402
-
-from popup import PanelWindow  # noqa: E402
+from gi.repository import GLib, Gtk
+from popup import PanelWindow
 
 # Long enough to read, short enough not to sit in the way of what you are doing.
 HIDE_AFTER_MS = 1400
@@ -44,7 +43,7 @@ def volume() -> tuple[int, bool]:
     out = run("wpctl", "get-volume", "@DEFAULT_AUDIO_SINK@")
     muted = "MUTED" in out
     try:
-        return int(round(float(out.split()[1]) * 100)), muted
+        return round(float(out.split()[1]) * 100), muted
     except (IndexError, ValueError):
         return 0, muted
 
@@ -53,7 +52,7 @@ def microphone() -> tuple[int, bool]:
     out = run("wpctl", "get-volume", "@DEFAULT_AUDIO_SOURCE@")
     muted = "MUTED" in out
     try:
-        return int(round(float(out.split()[1]) * 100)), muted
+        return round(float(out.split()[1]) * 100), muted
     except (IndexError, ValueError):
         return 0, muted
